config/ belongs under compute/
